Implement deform_conv2d for tch backend
Add native deformable convolution implementation for the LibTorch backend,
which was previously unimplemented due to tch-rs not exposing torchvision's
deform_conv2d bindings.

Implementation includes:
- Forward pass using deform_im2col + batched matrix multiplication
- Backward pass with gradients for input, offset, weight, mask, and bias
- Bilinear interpolation for sampling at fractional positions
- Bilinear splatting for gradient accumulation in backward pass

Also fixes a clippy warning in qtensor.rs (unnecessary cast).

+++ crates/burn-tch/src/ops/qtensor.rs
@@ -182,7 +182,7 @@ impl<E: TchElement> QTensorOps<Self> for LibTorch<E> {
         // Compute scale for symmetric quantization
         let (a, b) = scheme.value.range();
         let alpha = min_val.abs().max(max_val.abs());
-        let mut scale = (alpha + alpha) / (b as f32 - a as f32);
+        let mut scale = (alpha + alpha) / (b - a);
 
         // Avoid division by zero
         if scale == 0.0 {
